Remove definition_context from AutoNamed. Instead, pass reversed_context to resolve().
Rationale: definition_context did not provide any new functionality. It
was meant to support the definition of different parts of the model in
different functions without having to call model() separately for each
part. However this did not work in practice, apparently because the
locals dict of a function is emptied upon its return so references to
the locals dict did not provide the necessary context after the function
that created the variables returned.